Add detail to initial instructions, plus turn-warn beeps
In verbose mode, the first instruction we hear will include the name of
the road we should be on and the approximate direction we should travel.
(The same when we have got back on track.)

When approaching a turn we hear two short beeps as we approach and a
long one "at" the turn; we can then give the turn instruction a little
earlier. This should mean we are more likely to know what to do before
we reach a junction (e.g. while waiting at a red light). For now, we use
some DTMF tones because they are to hand.