Don't require movement while waiting for GPS to settle
When waiting for the GPS to settle we count a small number of updates
from the LocationManager before continuing - however we were requesting
to be updated only when we were moving. This meant we had to move some
distance before LiveRide guidance would start. This commit allows
LiveRideStates to request updates even when stationary; the
LiveRideStart and HuntForSegment states now do this.